add: refactor cms_users tabs and pages to avoid code duplication
